# implementations/dag-core/openapi/swagger.yaml
basePath: /
definitions:
  dto.ErrorResponse:
    properties:
      error:
        type: string
      run_id:
        type: string
      status:
        type: string
      symbol:
        type: string
    type: object
  events.StateDiffField:
    properties:
      after: {}
      before: {}
      field:
        type: string
    type: object
  handler.backtestRunRequest:
    properties:
      account_balance:
        type: number
      fee_bps:
        type: number
      from:
        type: string
      gap_handling:
        type: string
      min_lot:
        type: number
      mode:
        type: string
      parameter_set_id:
        type: string
      partition:
        type: string
      run_id:
        type: string
      slippage_bps:
        type: number
      source:
        type: string
      spread_bps:
        type: number
      symbol_code:
        type: string
      timeframe_code:
        type: string
      timezone:
        type: string
      to:
        type: string
      workflow_name:
        type: string
      workflow_version:
        type: string
    type: object
...
  response.BacktestMarketdataInput:
    properties:
      from:
        type: string
      gap_handling:
        type: string
      source:
        type: string
      symbol_code:
        type: string
      timeframe_code:
        type: string
      timezone:
        type: string
      to:
        type: string
    type: object
  response.BacktestRunResponse:
    properties:
      cycle_count:
        type: integer
      entrypoint:
        type: string
      marketdata:
        $ref: '#/definitions/response.BacktestMarketdataInput'
      mode:
        type: string
      parameter_set_id:
        type: string
      partition:
        type: string
      run_id:
        type: string
      status:
        type: string
      symbol:
        type: string
      window_size_bars:
        type: integer
      workflow_name:
        type: string
      workflow_version:
        type: string
    type: object
...
  response.ListRunStepsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/usecase.RunStepView'
        type: array
      run_id:
        type: string
    type: object
  response.ListRunsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/usecase.RunView'
        type: array
    type: object
...
  response.RunBacktestSummaryResponse:
    properties:
      summary:
        $ref: '#/definitions/usecase.BacktestSummaryView'
    type: object
  response.RunCompareResponse:
    properties:
      compare:
        $ref: '#/definitions/usecase.RunCompareView'
    type: object
  response.RunDetailResponse:
    properties:
      run:
        $ref: '#/definitions/usecase.RunView'
    type: object
  response.RunNodeDetailResponse:
    properties:
      node:
        $ref: '#/definitions/usecase.RunStepView'
      run_id:
        type: string
      sequence_no:
        type: string
    type: object
...
  usecase.BacktestSummaryView:
    properties:
      equity_point_count:
        type: integer
      has_strategy_fields:
        type: boolean
      max_drawdown:
        type: number
      partition:
        type: string
      run_id:
        type: string
      total_net_pnl:
        type: number
      trade_count:
        type: integer
      win_rate:
        type: number
    type: object
  usecase.RunCompareView:
    properties:
      base:
        $ref: '#/definitions/usecase.RunView'
      state_field_deltas:
        items:
          $ref: '#/definitions/usecase.RunStateFieldDeltaView'
        type: array
      summary_delta:
        $ref: '#/definitions/usecase.RunSummaryDeltaView'
      target:
        $ref: '#/definitions/usecase.RunView'
    type: object
  usecase.RunStateFieldDeltaView:
    properties:
      base_after: {}
      changed:
        type: boolean
      field:
        type: string
      target_after: {}
    type: object
  usecase.RunStepView:
    properties:
      duration_ns:
        type: integer
      error:
        type: string
      execution_id:
        type: string
      input_ref:
        type: string
      intent_id:
        type: string
      node_execution_id:
        type: string
      node_id:
        type: string
      node_name:
        type: string
      output_ref:
        type: string
      sequence_no:
        type: integer
      skip_reason:
        type: string
      snapshot_after_ref:
        type: string
      snapshot_before_ref:
        type: string
      state_diff:
        items:
          $ref: '#/definitions/events.StateDiffField'
        type: array
      status:
        type: string
      trade_id:
        type: string
      trigger_reason:
        type: string
    type: object
  usecase.RunSummaryDeltaView:
    properties:
      duration_ms_delta:
        type: integer
      failed_step_delta:
        type: integer
      retry_count_delta:
        type: integer
      skipped_step_delta:
        type: integer
      status_changed:
        type: boolean
      step_count_delta:
        type: integer
      succeeded_step_delta:
        type: integer
    type: object
  usecase.RunView:
    properties:
      duration_ms:
        type: integer
      ended_at:
        type: string
      error:
        type: string
      event_time:
        type: string
      event_type:
        type: string
      partition:
        type: string
      retry_count:
        type: integer
      run_id:
        type: string
      started_at:
        type: string
      status:
        type: string
      step_count:
        type: integer
    type: object
info:
  contact: {}
  description: DAG runtime HTTP API (auto-generated via swag).
  title: DAG Observatory API
  version: 0.1.0
paths:
  /algotrade/backtests:run:
    post:
      consumes:
      - application/json
      description: Triggers the configured runtime workflow as a backtest entrypoint
        over an explicit historical range and returns run correlation metadata.
      parameters:
      - description: Backtest run request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.backtestRunRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BacktestRunResponse'
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/response.BacktestRunResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Run real-data backtest workflow
      tags:
      - algotrade
...
  /runs:
    get:
      description: Returns run-level read model items. Optional `partition` narrows
        results.
      parameters:
      - description: Runtime partition filter
        in: query
        name: partition
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ListRunsResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List workflow runs
      tags:
      - dag
  /runs/{run_id}:
    get:
      description: Returns one run-level read model by run id.
      parameters:
      - description: Run ID
        in: path
        name: run_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RunDetailResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get run detail
      tags:
      - dag
  /runs/{run_id}/backtest-summary:
    get:
      description: Returns run-level backtest summary projection reconstructed from
        run-step state diff fields.
      parameters:
      - description: Run ID
        in: path
        name: run_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RunBacktestSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get run backtest summary
      tags:
      - dag
  /runs/{run_id}/steps:
    get:
      description: Returns node execution steps for one run.
      parameters:
      - description: Run ID
        in: path
        name: run_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ListRunStepsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List run steps
      tags:
      - dag
  /runs/{run_id}/steps/{sequence_no}:
    get:
      description: Returns one node execution record by run id and sequence number.
      parameters:
      - description: Run ID
        in: path
        name: run_id
        required: true
        type: string
      - description: Node execution sequence number
        in: path
        name: sequence_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RunNodeDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get run node execution detail
      tags:
      - dag
  /runs/compare:
    get:
      description: Returns summary deltas and state-field deltas between base and
        target run IDs.
      parameters:
      - description: Base run ID
        in: query
        name: base_run_id
        required: true
        type: string
      - description: Target run ID
        in: query
        name: target_run_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RunCompareResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Compare two workflow runs
      tags:
      - dag
...
swagger: "2.0"